+++ MazeRats/Public/Invoke-Spell.ps1
@@ -1,19 +1,4 @@
 function Invoke-Spell {
-    <#
-    .SYNOPSIS
-        Short description
-    .DESCRIPTION
-        Long description
-    .EXAMPLE
-        PS C:\> <example usage>
-        Explanation of what the example does
-    .INPUTS
-        Inputs (if any)
-    .OUTPUTS
-        Output (if any)
-    .NOTES
-        General notes
-    #>
     param (
         [MazeRat]
         $Player,